+++ baseplate/ratelimit/backends/__init__.py
@@ -1,0 +1,6 @@
+from .base import RateLimitBackend
+
+
+__all__ = [
+    'RateLimitBackend',
+]
